pyproject.toml
[project]
name = "monguite"
version = "0.1.0"
description = ""
authors = [{ name = "Felipe Vieira", email = "me@felipet.io" }]
requires-python = ">=3.10,<4"
license = "GPLv3"
dependencies = [
    "Django>=5.1,<6",
    "django-environ>=0.11.2,<0.12",
    "psycopg2-binary>=2.9.9,<3",
    "django-redis>=5.4.0,<6",
    "requests>=2.32.0,<3",
    "djangorestframework>=3.16.1,<4",
    "django-filter>=25.2,<26",
    "drf-spectacular>=0.29.0,<0.30",
    "gunicorn>=23.0.0,<24",
    "django-extensions>=3.2.3,<4",
    "django-mcp-server>=0.5.0",
    "httpx>=0.27.0,<0.28",
]

[dependency-groups]
dev = [
    "Werkzeug>=3.0.4,<4",
    "ipdb>=0.13.13,<0.14",
    "ruff>=0.11.6,<0.12",
    "coverage>=7.6.1,<8",
    "pylint-django>=2.5.5,<3",
    "pre-commit>=3.8.0,<4",
    "factory-boy>=3.3.1,<4",
    "django-debug-toolbar>=4.4.6,<5",
    "django-coverage-plugin>=3.1.0,<4",
    "pytest-django>=4.9.0,<5",
    "pytest-asyncio>=0.25.2,<0.26",
    "pytest-cov>=7.0.0",
    "pytest>=8.4.2",
]

[tool.uv]
package = false
default-groups = [
    "dev",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    "migrations",
    "static",
    "docs",
    "node_modules",
    "venv",
    ".venv",
    ".git",
    ".tox",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "DJ",  # flake8-django
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip without strict parameter
    "DJ001", # null=True on string fields is acceptable for external data integration
    "B904",  # exception chaining - not always necessary for Django management commands
    "DJ012", # method ordering - __str__ after save is acceptable
    "E722",  # bare except - acceptable in specific error handling cases
    "B017",  # assertRaises(Exception) - acceptable in tests for general error catching
]

[tool.ruff.lint.per-file-ignores]
"config/settings/*.py" = [
    "F403",  # star imports are standard for Django settings modules
    "F405",  # names from star imports are intentional in settings
]

[tool.ruff.lint.isort]
known-first-party = ["app", "config"]
section-order = ["future", "standard-library", "django", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django" = ["django"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
